Reader: Update action button link of an empty list of a user profile (#98706)
The issue was that on an empty list page, the "Back to Following" button with link `/read` was being displayed even when navigating from a user profile. This PR fixes this by updating the implementation to display a "Back to User Profile" button with link `/read/users/:id/lists` instead, allowing users to easily navigate back to the user profile.